+++ components/analytics/metric/api/src/main/java/com/flipperdevices/metric/api/events/complex/FlipperRPCInfoEvent.kt
@@ -7,15 +7,19 @@ data class FlipperRPCInfoEvent(
     val internalFreeBytes: Long,
     val internalTotalBytes: Long,
     val externalFreeBytes: Long,
-    val externalTotalBytes: Long
+    val externalTotalBytes: Long,
+    val firmwareForkName: String?,
+    val firmwareGitUrl: String?
 ) : ComplexEvent("flipper_rpc_info") {
-    override fun getParamsMap(): Map<String, Any> {
+    override fun getParamsMap(): Map<String, Any?> {
         return mapOf(
             "sdcard_is_available" to sdCardIsAvailable,
             "internal_free_byte" to internalFreeBytes,
             "internal_total_byte" to internalTotalBytes,
             "external_free_byte" to externalFreeBytes,
-            "external_total_byte" to externalTotalBytes
+            "external_total_byte" to externalTotalBytes,
+            "firmware_fork_name" to firmwareForkName,
+            "firmware_git_url" to firmwareGitUrl
         )
     }
 }