+++ components/analytics/metric/impl/src/main/java/com/flipperdevices/metric/impl/countly/CountlyApi.kt
@@ -1,5 +1,5 @@
 package com.flipperdevices.metric.impl.countly
 
 interface CountlyApi {
-    fun reportEvent(id: String, params: Map<String, Any>? = null)
+    fun reportEvent(id: String, params: Map<String, Any?>? = null)
 }