+++ components/bridge/api/src/main/java/com/flipperdevices/bridge/api/di/FlipperBleComponentInterface.kt
@@ -1,9 +1,0 @@
-package com.flipperdevices.bridge.api.di
-
-import android.bluetooth.BluetoothAdapter
-import com.flipperdevices.bridge.api.scanner.FlipperScanner
-
-interface FlipperBleComponentInterface {
-    val flipperScanner: FlipperScanner
-    val bluetoothAdapter: BluetoothAdapter
-}